# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_nurik_Dev_repos_mengu-backend_internal_email.AttachmentPayload:
    properties:
      content_type:
        example: application/pdf
        type: string
      filename:
        example: contract.pdf
        type: string
      size:
        example: 102400
        type: integer
      url:
        example: https://storage.example.com/contract.pdf
        type: string
    type: object
  github_com_nurik_Dev_repos_mengu-backend_internal_email.WebhookPayload:
    properties:
      attachments:
        items:
          $ref: '#/definitions/github_com_nurik_Dev_repos_mengu-backend_internal_email.AttachmentPayload'
        type: array
      body:
        example: We need to review the updated contract terms...
        type: string
      from:
        example: sender@company.com
        format: email
        type: string
      subject:
        example: Contract Review Meeting
        type: string
    type: object
...
  internal_tasks.taskListItem:
    properties:
      assignee_id:
        example: user_001
        type: string
      created_at:
        example: "2026-06-10T12:02:00Z"
        type: string
      description:
        example: Review the updated contract terms from partner
        type: string
      due_date:
        example: "2026-06-17T00:00:00Z"
        type: string
      id:
        example: task_001
        type: string
      status:
        enum:
        - pending
        - in_progress
        - completed
        - cancelled
        example: pending
        type: string
      title:
        example: Prepare contract review
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@mengu.ai
    name: API Support
  description: AI-assisted email automation platform. Ingests emails, analyzes them
    with LLM to extract intent and actions, then executes actions through deterministic
    handlers.
  license:
    name: Proprietary
  termsOfService: https://mengu.ai/terms
  title: Mengu AI API
  version: 1.0.0
paths:
...
  /tasks:
    get:
      consumes:
      - application/json
      description: List tasks for the authenticated user's organization with optional
        status filter.
      parameters:
      - description: Filter by status (pending, in_progress, completed, cancelled)
        in: query
        name: status
        type: string
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 20)
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/internal_tasks.taskListItem'
                type: array
              page:
                type: integer
              per_page:
                type: integer
              total:
                type: integer
            type: object
      security:
      - Bearer: []
      summary: List tasks
      tags:
      - Tasks
  /tasks/{id}:
    get:
      description: Get a single task by ID.
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              description:
                type: string
              due_date:
                type: string
              id:
                type: string
              status:
                type: string
              title:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
              message:
                type: string
            type: object
      security:
      - Bearer: []
      summary: Get task
      tags:
      - Tasks
    patch:
      consumes:
      - application/json
      description: Update a task's status or assignee.
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: request
        required: true
        schema:
          properties:
            assignee_id:
              type: string
            status:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              id:
                type: string
              status:
                type: string
              title:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
              message:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
              message:
                type: string
            type: object
      security:
      - Bearer: []
      summary: Update task
      tags:
      - Tasks
  /webhooks/email:
    post:
      consumes:
      - application/json
      description: Receive incoming email from external email service. Looks up organization
        by X-Webhook-Secret and stores as incoming_event.
      parameters:
      - description: Organization webhook secret
        in: header
        name: X-Webhook-Secret
        required: true
        type: string
      - description: Email payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_nurik_Dev_repos_mengu-backend_internal_email.WebhookPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              event_id:
                type: string
              status:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
              message:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
              message:
                type: string
            type: object
      security:
      - WebhookSecret: []
      summary: Ingest email via webhook
      tags:
      - Webhooks
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and the JWT access token.
    in: header
    name: Authorization
    type: apiKey
  WebhookSecret:
    description: Webhook secret key for authenticating email webhook requests.
    in: header
    name: X-Webhook-Secret
    type: apiKey
swagger: "2.0"